backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.94.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mem0ai>=1.0.11",
    "openai>=2.31.0",
    "bcrypt>=4.0.0",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
    "psycopg2-binary>=2.9.11",
]
